requirements.txt
pandas==1.3.5; python_version == '3.7'
pandas==2.0.1; python_version > '3.7'
google-cloud-aiplatform[pipelines]==1.26.0
google-auth==2.17.3
discord
python-dotenv
